+++ src/main/scala/model/entities/Enemy.scala
@@ -1,6 +1,8 @@
 package model.entities
 
+import model.entities.WorldSpace.Position
+
 trait Enemy extends MovingEntity with AttackingEntity:
-  override def velocity: Double = 1.0
+  override def velocity: Float = 1.0
 
 class Zombie() extends Enemy
